pyproject.toml
[project]
name = "personal-assistant-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "google-generativeai>=0.8.5",
    "llama-index>=0.12.0,<0.13.0",
    "llama-index-llms-gemini>=0.5.0,<0.6.0",
    "llama-index-agent-openai>=0.4.12",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "telethon>=1.40.0",
]
